fix: 포메이션 공정조건 전압/전류 컬럼 타입 smallint → float 변경

+++ src/common/entities/worklogs/worklog-13-formation.entity.ts
@@ -264,13 +264,13 @@ export class WorklogFormation extends WorklogBase {
   @Column({ type: 'varchar', length: 50, nullable: true })
   preFormationVoltageCondition: string;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   preFormationLowerVoltage: number;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   preFormationUpperVoltage: number;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   preFormationAppliedCurrent: number;
 
   @Column({ type: 'smallint', nullable: true })
@@ -280,13 +280,13 @@ export class WorklogFormation extends WorklogBase {
   @Column({ type: 'varchar', length: 50, nullable: true })
   mainFormationVoltageCondition: string;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   mainFormationLowerVoltage: number;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   mainFormationUpperVoltage: number;
 
-  @Column({ type: 'smallint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   mainFormationAppliedCurrent: number;
 
   @Column({ type: 'smallint', nullable: true })
